Move Tests
Moved all `pdflatex` tests to the tex submodule.
